requirements.txt
asgiref==3.9.1
asyncpg==0.30.0
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
schedule==1.2.2
sqlparse==0.5.3
tzdata==2025.2